fix: timeline() assumed 64-bit time_t — breaks the CLI build on armv7
Seven E0277/E0308 errors in libawdl-cli on the Pi 400, all in timeline(): it read
pkt.header.ts.tv_sec directly into i64 arithmetic, and tv_sec is time_t, which is
i32 on 32-bit targets. The other two capture-reading paths (frame 1764, 1849)
already cast tv_sec as u64 and were fine; timeline was the one that assumed the
width.

Normalised with `as i64` at the source, so every downstream calculation stays
i64 on both widths without scattering casts. Same class of bug as the tun.rs
fix, same reason it was invisible until a 32-bit Pi appeared.

The awdl binary now builds and runs on the Pi 400 (armv7l, Raspbian 11).

+++ crates/libawdl-cli/src/main.rs
@@ -209,7 +209,10 @@ fn timeline<T: pcap::Activated + ?Sized>(mut cap: pcap::Capture<T>, bucket_s: i6
     let mut last_bucket = 0i64;
 
     while let Ok(pkt) = cap.next_packet() {
-        let ts = pkt.header.ts.tv_sec;
+        // `as i64`: pcap's tv_sec is time_t, which is i32 on 32-bit targets (armv7, the Pi
+        // 400) and i64 elsewhere. Normalising here keeps every downstream calculation i64
+        // on both widths instead of scattering casts through the arithmetic below.
+        let ts = pkt.header.ts.tv_sec as i64;
         let base = *first_ts.get_or_insert(ts);
         let bucket = (ts - base) / bucket_s;
         last_bucket = last_bucket.max(bucket);
